feat: Refactor project structure for localization support
