+++ src/app/[locale]/page.tsx
@@ -1,10 +1,13 @@
 'use client';
 
 import firebaseAuthService from '@/services/firebase-service/firebase-auth-service';
+import { useTranslations } from 'next-intl';
 
 import { useEffect } from 'react';
 
 const Home = (): React.ReactNode => {
+  const t = useTranslations('HomePage');
+
   useEffect(() => {
     const handleCookieCheck = async (): Promise<void> => {
       const cookie = document.cookie.replace(/(?:(?:^|.*;\s*)identity\s*=\s*([^;]*).*$)|^.*$/, '');
@@ -15,7 +18,7 @@ const Home = (): React.ReactNode => {
     handleCookieCheck();
   }, []);
 
-  return <div>Home Page, show around 10 posts for the most active stocks of the day</div>;
+  return <div> {t('title')}</div>;
 };
 
 export default Home;
