+++ src/app/[locale]/post/[id]/page.tsx
