+++ src/app/[locale]/stocks/[id]/page.tsx
